Every product has its own threshold now, app asks for confirmation when try to delete a product, stock change type now chosen via radiobuttons, not combobox

+++ Project_Media_Bazaar/ManagementForm.cs
@@ -17,7 +17,7 @@ public partial class ManagementForm : Form
         lvProducts.Items.Clear();
         foreach (Product product in company.GetProducts())
         {
-            if (product.Amount > 50)
+            if (product.Amount > product.Threshold)
             {
                 string[] row = { product.Name, product.Amount.ToString(), product.Category.ToString() };
                 lvProducts.Items.Add(product.Id.ToString()).SubItems.AddRange(row);
@@ -90,8 +90,13 @@ public partial class ManagementForm : Form
     {
         try
         {
-            company.DeleteProduct(company.GetProducts()[lvProducts.FocusedItem.Index]);
-            RefreshListbox();
+            var confirmResult = MessageBox.Show("Are you sure to delete this item ?",
+                                     "Confirm Delete!", MessageBoxButtons.YesNo);
+            if(confirmResult == DialogResult.Yes)
+            {
+                company.DeleteProduct(company.GetProducts()[lvProducts.FocusedItem.Index]);
+                RefreshListbox();
+            }
         }
         catch (NullReferenceException)
         {
